+++ Changelog.md
@@ -9,6 +9,7 @@ Bugfixes:
  * Type Checker: Properly check array length and don't rely on an assertion in code generation.
  * Type Checker: Properly support overwriting members inherited from ``address`` in a contract
    (such as ``balance``, ``transfer``, etc.)
+ * Type Checker: Prevent duplicate event declarations.
 
 ### 0.4.17 (2017-09-21)